Refactor types for ShoppingCartManager
